[James] refactoring the service for suburb suggestions
